+++ x/marker/keeper/keeper_test.go
@@ -18,6 +18,7 @@ import (
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	"github.com/cosmos/cosmos-sdk/x/feegrant"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	"github.com/cosmos/cosmos-sdk/x/group"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	"github.com/cosmos/cosmos-sdk/x/quarantine"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -755,9 +756,28 @@ func TestCanForceTransferFrom(t *testing.T) {
 		app.AccountKeeper.SetAccount(ctx, acc)
 	}
 
+	createGroup := func() sdk.AccAddress {
+		goCtx := sdk.WrapSDKContext(ctx)
+		msg, err := group.NewMsgCreateGroupWithPolicy("cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn",
+			[]group.MemberRequest{
+				{
+					Address:  "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn",
+					Weight:   "1",
+					Metadata: "",
+				},
+			},
+			"", "", true, group.NewPercentageDecisionPolicy("0.5", time.Second, time.Second))
+		require.NoError(t, err, "NewMsgCreateGroupWithPolicy")
+		res, err := app.GroupKeeper.CreateGroupWithPolicy(goCtx, msg)
+		require.NoError(t, err, "CreateGroupWithPolicy")
+
+		return sdk.MustAccAddressFromBech32(res.GroupPolicyAddress)
+	}
+
 	addrNoAcc := sdk.AccAddress("addrNoAcc___________")
 	addrSeq0 := sdk.AccAddress("addrSeq0____________")
 	addrSeq1 := sdk.AccAddress("addrSeq1____________")
+	addrGroup := createGroup()
 	setAcc(addrSeq0, 0)
 	setAcc(addrSeq1, 1)
 
@@ -769,6 +789,7 @@ func TestCanForceTransferFrom(t *testing.T) {
 		{name: "address without an account", from: addrNoAcc, exp: true},
 		{name: "address with sequence 0", from: addrSeq0, exp: false},
 		{name: "address with sequence 1", from: addrSeq1, exp: true},
+		{name: "group address", from: addrGroup, exp: true},
 	}
 
 	for _, tc := range tests {
@@ -1943,7 +1964,7 @@ func TestBypassAddrsLocked(t *testing.T) {
 		sdk.AccAddress("addrs[4]____________"),
 	}
 
-	mk := markerkeeper.NewKeeper(nil, nil, paramtypes.NewSubspace(nil, nil, nil, nil, "test"), nil, &dummyBankKeeper{}, nil, nil, nil, nil, nil, addrs)
+	mk := markerkeeper.NewKeeper(nil, nil, paramtypes.NewSubspace(nil, nil, nil, nil, "test"), nil, &dummyBankKeeper{}, nil, nil, nil, nil, nil, addrs, nil)
 
 	// Now that the keeper has been created using the provided addresses, change the first byte of
 	// the first address to something else. Then, get the addresses back from the keeper and make